Eliminate narrowing in sgemm bench

+++ test/level3/sgemm/bench.cpp
@@ -56,15 +56,15 @@ static void gemm_arguments_increasing(benchmark::internal::Benchmark *b) {
   size_t minMem = 0;
   size_t maxMem = DRAM;
 
-  size_t m_multiple = 32;
-  size_t k_multiple = 96;
-  size_t n_multiple = 24;
-  for (size_t i = 1; i < 1000; i += 4) {
-    size_t M = i * m_multiple;
-    size_t N = i * n_multiple;
-    size_t K = i * k_multiple;
+  int m_multiple = 32;
+  int k_multiple = 96;
+  int n_multiple = 24;
+  for (int i = 1; i < 1000; i += 4) {
+    int M = i * m_multiple;
+    int N = i * n_multiple;
+    int K = i * k_multiple;
 
-    size_t total = M * N + M * K + N * K;
+    size_t total = (size_t)M * N + (size_t)M * K + (size_t)N * K;
     total *= 4;
 
     if (total < maxMem && total >= minMem) {
